backend/requirements.txt
# FastAPI framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
alembic==1.13.3

# AWS SDK
boto3==1.35.30
botocore==1.35.30

# Image processing
Pillow==10.4.0
python-multipart==0.0.12
ImageHash==4.3.1

# Testing
pytest==8.3.0
pytest-cov==5.0.0
pytest-asyncio==0.24.0
httpx==0.27.0

# Code quality
black==24.8.0
pylint==3.3.0
mypy==1.11.0

# Other
python-dotenv==1.0.1
redis==5.1.0
